# docs/api/petlife-openapi.yaml
openapi: 3.0.3
info:
  title: PetLife 宠物生活管家 API
  description: |
    # PetLife 宠物生活管家完整 API 文档
    
    ## 项目概述
    - **项目代号**: petlife
    - **产品名称**: 宠物生活管家
    - **交付范围**: mobile-app + server + admin-web
    
    ## 技术规范
    
    ### 鉴权约定
    - Header: `Authorization: Bearer <access_token>`
    - refresh token 通过专用接口换取
    
    ### 字段命名
    - 统一使用 `snake_case`
    
    ### 时间格式
    - ISO 8601 (UTC)
    
    ### 响应结构
    ```json
    {
      "code": "OK",
      "message": "success",
      "data": {}
    }
    ```
    
  version: 1.0.0
  contact:
    name: PetLife Team
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
  - url: https://api-dev.petlife.com/api/v1
    description: 开发环境
  - url: https://api.petlife.com/api/v1
    description: 生产环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 使用 access_token 进行鉴权
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: string
          example: "OK"
          description: 响应码
        message:
          type: string
          example: "success"
          description: 响应消息
        data:
          type: object
          description: 响应数据
...
    AmapGeocodeResult:
      type: object
      properties:
        matched:
          type: boolean
        address:
          type: string
        formatted_address:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        city_code:
          type: string
          nullable: true
        adcode:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        level:
          type: string
          nullable: true
    AmapReverseGeocodeResult:
      type: object
      properties:
        matched:
          type: boolean
        formatted_address:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        township:
          type: string
          nullable: true
        city_code:
          type: string
          nullable: true
        adcode:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
...
security:
  - BearerAuth: []
paths:
...
  /services/home:
    get:
      tags:
        - 服务 (Service)
      summary: 服务首页聚合
      description: |
        获取服务首页聚合信息
        
        **返回内容**:
        - city_code, city_name
        - opened (是否开通)
        - unavailable_reason (未开通原因)
        - categories (服务分类)
        - featured_providers (推荐服务商)
        - upcoming_appointments (即将到来的预约)
        - commerce_placeholder (商城占位)
        
        **说明**:
        - commerce_placeholder 只承接商城当前预留说明，不进入交易后端链路
        - 当前城市没有在线服务商时，接口返回未开通态，而不是空数组空白页
      parameters:
        - name: pet_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: 宠物 ID
        - name: city_code
          in: query
          required: false
          schema:
            type: string
          description: 城市编码
      responses:
        '200':
          description: 获取成功
  /providers:
    get:
      tags:
        - 服务 (Service)
      summary: 获取服务商列表
      description: |
        获取服务商列表
        
        **说明**:
        - provider_type 支持 hospital、boarding、grooming、training
        - 响应会返回服务项目和近期可预约时段摘要
        - 传入 latitude + longitude 时返回 distance_meters；sort=distance 时按距离升序排序
      parameters:
        - name: provider_type
          in: query
          required: false
          schema:
            type: string
            enum: [hospital, boarding, grooming, training]
          description: 服务商类型
        - name: city_code
          in: query
          required: false
          schema:
            type: string
          description: 城市编码
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [distance, rating]
          description: 排序方式
        - name: latitude
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          description: 用户当前位置纬度
        - name: longitude
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          description: 用户当前位置经度
      responses:
        '200':
          description: 获取成功；服务商响应新增 coordinate_source 与 distance_meters（未传位置时为 null）
  /providers/{providerId}:
    get:
      tags:
        - 服务 (Service)
      summary: 获取服务商详情
      description: 获取服务商详细信息
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 服务商 ID
      responses:
        '200':
          description: 获取成功
  /providers/{providerId}/slots:
    get:
      tags:
        - 服务 (Service)
      summary: 获取服务商可预约时段
      description: |
        获取服务商的可预约时段
        
        **说明**:
        - 时段名额由服务端读取 provider_schedule_slots
        - available_quota = quota - booked_count
        - 只有 status=open 且剩余名额大于 0 时 bookable=true
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 服务商 ID
        - name: appointment_type
          in: query
          required: false
          schema:
            type: string
            enum: [hospital, boarding, grooming, training]
          description: 预约类型
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: 开始日期
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: 结束日期
      responses:
        '200':
          description: 获取成功
  /appointments:
    get:
      tags:
        - 服务 (Service)
      summary: 获取预约记录
      description: |
        获取用户的预约记录
        
        **状态筛选**:
        - all: 全部
        - pending_confirm: 待确认
        - confirmed: 已确认
        - completed: 已完成
        - canceled: 已取消
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, pending_confirm, confirmed, completed, canceled]
          description: 状态筛选
      responses:
        '200':
          description: 获取成功
          
    post:
      tags:
        - 服务 (Service)
      summary: 创建预约
      description: |
        创建服务预约
        
        **说明**:
        - 创建预约会校验当前用户是否有该宠物访问权
        - 创建预约会校验服务商状态、预约类型和时段名额
        - 同一时段通过数据库行锁和名额递增防止超卖
        - 预约创建后会生成站内预约通知，并派生宠物时间轴 service 事件
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pet_id
                - provider_id
                - appointment_type
                - appointment_date
                - appointment_slot
                - contact_name
                - contact_mobile
              properties:
                pet_id:
                  type: string
                  example: "10001"
                  description: 宠物 ID
                provider_id:
                  type: string
                  example: "70001"
                  description: 服务商 ID
                appointment_type:
                  type: string
                  example: "hospital"
                  enum: [hospital, boarding, grooming, training]
                  description: 预约类型
                appointment_date:
                  type: string
                  format: date
                  example: "2026-04-22"
                  description: 预约日期
                appointment_slot:
                  type: string
                  example: "10:00-11:00"
                  description: 预约时段
                demand_desc:
                  type: string
                  example: "最近轻微拉稀，想检查"
                  description: 需求描述
                contact_name:
                  type: string
                  example: "李明"
                  description: 联系人姓名
                contact_mobile:
                  type: string
                  example: "13800000000"
                  description: 联系电话
      responses:
        '200':
          description: 创建成功
  /appointments/{appointmentId}/cancel:
    patch:
      tags:
        - 服务 (Service)
      summary: 取消预约
      description: |
        取消预约
        
        **说明**:
        - 仅 pending_confirm 和 confirmed 状态允许取消
        - 取消后会回补原时段名额，并更新时间轴事件为取消态
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 预约 ID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                cancel_reason:
                  type: string
                  example: "临时改期"
                  description: 取消原因
      responses:
        '200':
          description: 取消成功

          
  # ==================== 通知模块 ====================
...
  /admin/map/geocode:
    get:
      tags:
        - 后台地图 (Admin Map)
      summary: 高德地理编码辅助
      description: 根据地址调用高德 Web 服务返回经纬度候选，供后台维护服务商坐标。
      parameters:
        - name: address
          in: query
          required: true
          schema:
            type: string
            maxLength: 255
          description: 待解析地址
        - name: city
          in: query
          required: false
          schema:
            type: string
            maxLength: 32
          description: 城市编码或城市名
      responses:
        '200':
          description: 解析成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AmapGeocodeResult'
        '503':
          description: 高德 Web 服务 Key 未配置
  /admin/map/reverse-geocode:
    get:
      tags:
        - 后台地图 (Admin Map)
      summary: 高德逆地理编码辅助
      description: 根据经纬度调用高德 Web 服务返回地址信息，供后台核对服务商坐标。
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          description: 纬度
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          description: 经度
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AmapReverseGeocodeResult'
        '503':
          description: 高德 Web 服务 Key 未配置
...
  /admin/service/providers:
    get:
      tags:
        - 后台服务 (Admin Service)
      summary: 获取服务商列表（管理端）
      description: 获取服务商列表，支持更多筛选条件
      parameters:
        - name: provider_type
          in: query
          required: false
          schema:
            type: string
            enum: [hospital, boarding, grooming, training]
          description: 服务商类型
        - name: city_code
          in: query
          required: false
          schema:
            type: string
          description: 城市编码
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [online, rest, offline]
          description: 状态筛选
      responses:
        '200':
          description: 获取成功
          
    post:
      tags:
        - 后台服务 (Admin Service)
      summary: 创建服务商（管理端）
      description: 创建新的服务商
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider_type
                - provider_name
                - city_code
                - status
              properties:
                provider_type:
                  type: string
                  example: "hospital"
                  enum: [hospital, boarding, grooming, training]
                  description: 服务商类型
                provider_name:
                  type: string
                  example: "爱宠宠物医院"
                  description: 服务商名称
                city_code:
                  type: string
                  example: "310000"
                  description: 城市编码
                address:
                  type: string
                  example: "上海市浦东新区XX路XX号"
                  description: 地址
                latitude:
                  type: number
                  format: double
                  example: 31.2304
                  description: 纬度
                longitude:
                  type: number
                  format: double
                  example: 121.4737
                  description: 经度
                coordinate_source:
                  type: string
                  enum: [manual, amap]
                  example: "amap"
                  description: 坐标来源
                contact_phone:
                  type: string
                  example: "021-12345678"
                  description: 联系电话
                business_hours:
                  type: string
                  example: "09:00-18:00"
                  description: 营业时间
                rating_avg:
                  type: number
                  format: double
                  example: 4.8
                  description: 平均评分
                review_count:
                  type: integer
                  example: 36
                  description: 评价数量
                status:
                  type: string
                  example: "online"
                  enum: [online, rest, offline]
                  description: 服务商状态
      responses:
        '200':
          description: 创建成功
  /admin/service/providers/{providerId}:
    patch:
      tags:
        - 后台服务 (Admin Service)
      summary: 更新服务商（管理端）
      description: 更新服务商信息
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 服务商 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider_name:
                  type: string
                  description: 服务商名称
                address:
                  type: string
                  description: 地址
                contact_phone:
                  type: string
                  description: 联系电话
                business_hours:
                  type: string
                  description: 营业时间
                status:
                  type: string
                  enum: [online, rest, offline]
                  description: 状态
      responses:
        '200':
          description: 更新成功
...
  /admin/service/providers/{providerId}/slots:
    post:
      tags:
        - 后台服务 (Admin Service)
      summary: 创建预约时段（管理端）
      description: 为服务商创建预约时段
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 服务商 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - appointment_type
                - slot_date
                - start_time
                - end_time
                - quota
                - status
              properties:
                appointment_type:
                  type: string
                  example: "hospital"
                  enum: [hospital, boarding, grooming, training]
                  description: 预约类型
                slot_date:
                  type: string
                  format: date
                  example: "2026-05-01"
                  description: 时段日期
                start_time:
                  type: string
                  example: "10:00:00"
                  description: 开始时间
                end_time:
                  type: string
                  example: "11:00:00"
                  description: 结束时间
                quota:
                  type: integer
                  example: 5
                  description: 名额
                status:
                  type: string
                  example: "open"
                  enum: [open, closed]
                  description: 状态
      responses:
        '200':
          description: 创建成功
...
  /admin/service/appointments:
    get:
      tags:
        - 后台服务 (Admin Service)
      summary: 获取预约列表（管理端）
      description: 获取预约记录列表，支持更多筛选条件
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, pending_confirm, confirmed, completed, canceled]
          description: 状态筛选
        - name: provider_type
          in: query
          required: false
          schema:
            type: string
            enum: [hospital, boarding, grooming, training]
          description: 服务商类型
        - name: city_code
          in: query
          required: false
          schema:
            type: string
          description: 城市编码
      responses:
        '200':
          description: 获取成功
...